client: split client, use ruff
