+++ ollama_x/api/exceptions.py
@@ -39,7 +39,6 @@ class APIError(BaseModel, Generic[TE]):
 
 
 class BaseAPIException(Exception):
-
     status_code: int = 500
 
     def __init__(self, detail: str) -> None:
