+++ ollama_x/client/inner.py
@@ -1,7 +1,5 @@
 import asyncio
 import json
-import ssl
-from contextlib import asynccontextmanager
 from functools import partial
 from typing import Any, AsyncIterable, Callable
 
@@ -17,113 +15,7 @@ from ollama_x.api import endpoints
 from ollama_x.types import OllamaModel, OpenAIModel
 
 
-def get_model_params(model_name: str) -> tuple[str, str]:
-    version = "latest"
-
-    if "/" not in model_name:
-        model_name = f"library/{model_name}"
-
-    if ":" in model_name:
-        model_name, version = model_name.split(":", 1)
-
-    return model_name, version
-
-
-class OllamaClient:
-    """Client to Ollama Server API."""
-
-    def __init__(self, base_url: str | HttpUrl | None, is_self: bool = False):
-        self.base_url = str(base_url) if base_url is not None else None
-        self.is_self = is_self
-
-    @asynccontextmanager
-    async def send_request(
-        self,
-        path: str,
-        method: str,
-        base_url: None | str = None,
-        **kwargs,
-    ) -> aiohttp.ClientResponse:
-        """Send request to the server."""
-
-        ssl_context = ssl.create_default_context(cafile=certifi.where())
-
-        if self.base_url is None and base_url is None:
-            raise RuntimeError("Base url is not provided.")
-
-        async with aiohttp.ClientSession(
-            base_url or self.base_url, connector=TCPConnector(ssl=ssl_context)
-        ) as session:
-            async with getattr(session, method)(path, timeout=5, **kwargs) as response:
-                yield response
-
-    def list_models(self) -> aiohttp.ClientResponse:
-        """List all models available on the server."""
-
-        return self.send_request("/api/tags", "get")
-
-    def show_model_info(self, model_name: str, verbose=False) -> aiohttp.ClientResponse:
-        """Show model information."""
-
-        return self.send_request(
-            f"/api/show",
-            "post",
-            json={
-                "name": model_name,
-                "verbose": verbose,
-            },
-        )
-
-    def list_running_models(self) -> aiohttp.ClientResponse:
-        """List all running models."""
-
-        return self.send_request("/api/ps", "get")
-
-    async def tokenize(self, model_name: str, prompt: str) -> aiohttp.ClientResponse:
-        return self.send_request(
-            "/api/tokenize",
-            "post",
-            json={
-                "model": model_name,
-                "prompt": prompt,
-            },
-        )
-
-    async def get_manifest(self, model_name: str):
-
-        model_name, version = get_model_params(model_name)
-
-        return await self.send_request(
-            f"/v2/{model_name}/manifests/{version}",
-            "get",
-            base_url=f"https://registry.ollama.ai",
-        )
-
-    async def download_blob(self, model_name: str, digest: str) -> aiohttp.ClientResponse:
-        model_name, _ = get_model_params(model_name)
-
-        return await self.send_request(
-            f"/v2/{model_name}/blobs/{digest}",
-            "get",
-            base_url=f"https://registry.ollama.ai",
-        )
-
-    def chat(
-        self,
-        request: dict[str, Any],
-        headers: dict[str, Any],
-    ) -> aiohttp.ClientResponse:
-
-        return self.send_request(
-            "/api/chat",
-            "post",
-            json=request,
-            headers=headers,
-        )
-
-
 class ResponseIterable(AsyncIterable):
-
     EMPTY = object()
 
     def __init__(self):
@@ -159,12 +51,8 @@ class ResponseIterable(AsyncIterable):
             await asyncio.sleep(0)
 
 
-def run_in_thread(coro):
-    return asyncio.create_task(coro)
-
-
 class InnerClient:
-    def __init__(self, app: "FastAPI"):
+    def __init__(self, app: FastAPI):
         self.app = app
 
     async def send_request(
@@ -217,7 +105,7 @@ class InnerClient:
 
         loop = asyncio.get_running_loop()
         loop.call_soon(
-            run_in_thread,
+            asyncio.create_task,
             self.app(
                 {
                     "type": "http",
